requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
cryptography==46.0.3
distro==1.9.0
frozenlist==1.8.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
packaging==25.0
phonenumbers==9.0.22
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
redis==7.1.0
signalbot==0.21.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
websockets==16.0
yarl==1.22.0
